Split ImagingRequestView into separate files
