+++ packages/desktop/app/contexts/ElectronProvider.js
@@ -11,6 +11,7 @@ export const ElectronProvider = ({ children }) => {
   const showOpenDialog = (...args) => remote.dialog.showOpenDialog(...args);
   const showSaveDialog = (...args) => remote.dialog.showSaveDialog(...args);
   const openPath = path => shell.openPath(path);
+  const openUrl = url => shell.openExternal(url);
 
   return (
     <ElectronContext.Provider
@@ -18,6 +19,7 @@ export const ElectronProvider = ({ children }) => {
         showOpenDialog,
         showSaveDialog,
         openPath,
+        openUrl,
         printPage,
         readFile,
       }}
