+++ packages/desktop/app/views/patients/imagingRequest/index.js
@@ -1,0 +1,1 @@
+export * from './ImagingRequestView';
