+++ packages/desktop/app/views/patients/index.js
@@ -2,7 +2,7 @@ export { PatientListingView, AdmittedPatientsView, OutpatientsView } from './Pat
 export { PatientView } from './PatientView';
 export { EncounterView } from './EncounterView';
 export { LabRequestView } from './LabRequestView';
-export { ImagingRequestView } from './ImagingRequestView';
+export { ImagingRequestView } from './imagingRequest';
 export { DischargeSummaryView } from './DischargeSummaryView';
 export { ImmunisationsView } from './ImmunisationsView';
 export { CovidCampaignView } from './CovidCampaignView';
